Added indirect.c

+++ dft/ct-dif.c
@@ -18,7 +18,7 @@
  *
  */
 
-/* $Id: ct-dif.c,v 1.1 2002-06-08 13:34:58 athena Exp $ */
+/* $Id: ct-dif.c,v 1.2 2002-06-08 15:10:44 athena Exp $ */
 
 /* decimation in time Cooley-Tukey */
 #include "dft.h"
@@ -54,7 +54,7 @@ static int applicable(const solver_ct *ego, const problem *p_)
 		  && p->ri == p->ro
 
 		  /* if hardwired strides, test whether they match */
-		  && (!e->is || e->is == (d[0].n / e->radix) * d[0].is)
+		  && (!e->is || e->is == (int)(d[0].n / e->radix) * d[0].is)
 	       );
      }
      return 0;
